webpack for splunk app
