+++ app/onepassword_events_api/appserver/static/javascript/setup_page.js
@@ -1,0 +1,10 @@
+"use strict";
+
+import React from "react";
+import ReactDOM from "react-dom";
+import SetupPage from "./views/app";
+
+ReactDOM.render(
+	React.createElement(SetupPage),
+  	document.getElementById("main_container")
+);
