+++ app/onepassword_events_api/default/data/ui/views/setup_page_dashboard.xml
@@ -7,7 +7,7 @@ http://docs.splunk.com/Documentation/Splunk/latest/Viz/PanelreferenceforSimplifi
 -->
 <dashboard isDashboard="false"
            hideTitle="true"
-           script="javascript/setup_page_130.js"
+           script="build/main.js"
            stylesheet="styles/setup_page.css">
     <row>
         <panel>
