+++ app/onepassword_events_api/webpack.config.js
@@ -1,0 +1,9 @@
+const path = require("path");
+
+module.exports = {
+    entry: "./appserver/static/javascript/setup_page.js",
+    output: {
+        filename: "main.js",
+        path: path.resolve(__dirname, "appserver/static/build"),
+    },
+};
